requirements.txt
playwright>=1.40
requests>=2.31
supabase>=2.0
python-dotenv>=1.0
beautifulsoup4>=4.12
lxml>=5.0
